+++ CopilotKit/packages/react-core/src/hooks/use-tree.ts
@@ -1,4 +1,4 @@
-import { randomId } from "@copilotkit/shared";
+import { randomId } from "@bigppwong/copilotkit-shared";
 import { useCallback, useReducer } from "react";
 
 export type TreeNodeId = string;