+++ CopilotKit/packages/react-core/src/utils/extract.ts
@@ -4,7 +4,7 @@ import {
   MappedParameterTypes,
   Parameter,
   actionParametersToJsonSchema,
-} from "@copilotkit/shared";
+} from "@bigppwong/copilotkit-shared";
 import {
   ActionExecutionMessage,
   Message,
@@ -13,14 +13,14 @@ import {
   convertGqlOutputToMessages,
   CopilotRequestType,
   ForwardedParametersInput,
-} from "@copilotkit/runtime-client-gql";
+} from "@bigppwong/copilotkit-runtime-client-gql";
 import { CopilotContextParams, CopilotMessagesContextParams } from "../context";
 import { defaultCopilotContextCategories } from "../components";
-import { CopilotRuntimeClient } from "@copilotkit/runtime-client-gql";
+import { CopilotRuntimeClient } from "@bigppwong/copilotkit-runtime-client-gql";
 import {
   convertMessagesToGqlInput,
   filterAgentStateMessages,
-} from "@copilotkit/runtime-client-gql";
+} from "@bigppwong/copilotkit-runtime-client-gql";
 
 interface InitialState<T extends Parameter[] | [] = []> {
   status: "initial";