+++ CopilotKit/packages/react-ui/src/components/chat/Popup.tsx
@@ -9,16 +9,16 @@
  *
  * ## Install Dependencies
  *
- * This component is part of the [@copilotkit/react-ui](https://npmjs.com/package/@copilotkit/react-ui) package.
+ * This component is part of the [@bigppwong/copilotkit-react-ui](https://npmjs.com/package/@bigppwong/copilotkit-react-ui) package.
  *
- * ```shell npm2yarn \"@copilotkit/react-ui"\
- * npm install @copilotkit/react-core @copilotkit/react-ui
+ * ```shell npm2yarn \"@bigppwong/copilotkit-react-ui"\
+ * npm install @bigppwong/copilotkit-react-core @bigppwong/copilotkit-react-ui
  * ```
  * ## Usage
  *
  * ```tsx
- * import { CopilotPopup } from "@copilotkit/react-ui";
- * import "@copilotkit/react-ui/styles.css";
+ * import { CopilotPopup } from "@bigppwong/copilotkit-react-ui";
+ * import "@bigppwong/copilotkit-react-ui/styles.css";
  *
  * <CopilotPopup
  *   labels={{
@@ -33,7 +33,7 @@
  * By default, CopilotKit components do not have any styles. You can import CopilotKit's stylesheet at the root of your project:
  * ```tsx title="YourRootComponent.tsx"
  * ...
- * import "@copilotkit/react-ui/styles.css"; // [!code highlight]
+ * import "@bigppwong/copilotkit-react-ui/styles.css"; // [!code highlight]
  *
  * export function YourRootComponent() {
  *   return (