+++ CopilotKit/packages/runtime/src/lib/runtime/copilot-runtime.ts
@@ -6,7 +6,7 @@
  * ## Usage
  *
  * ```tsx
- * import { CopilotRuntime } from "@copilotkit/runtime";
+ * import { CopilotRuntime } from "@bigppwong/copilotkit-runtime";
  *
  * const copilotKit = new CopilotRuntime();
  * ```
@@ -23,7 +23,7 @@ import {
   CopilotKitLowLevelError,
   CopilotKitAgentDiscoveryError,
   CopilotKitMisuseError,
-} from "@copilotkit/shared";
+} from "@bigppwong/copilotkit-shared";
 import {
   CopilotServiceAdapter,
   EmptyAdapter,
@@ -239,7 +239,7 @@ export interface CopilotRuntimeConstructorParams<T extends Parameter[] | [] = []
   /**
    * A function that creates an MCP client instance for a given endpoint configuration.
    * This function is responsible for using the appropriate MCP client library
-   * (e.g., `@copilotkit/runtime`, `ai`) to establish a connection.
+   * (e.g., `@bigppwong/copilotkit-runtime`, `ai`) to establish a connection.
    * Required if `mcpServers` is provided.
    *
    * ```typescript